requirements.txt
gradio
torch
torchvision  # Often needed with torch
opencv-python
numpy
pillow